refactor(currency): move decimal rational helpers
